feat: code refactor indexDB
